python-backend/requirements.txt
fastapi[standard]
supabase
twilio
google-generativeai
firebase-admin
httpx[http2]
uvicorn[standard]